# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SignalWatch – Healthcare Social Listening API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List sources for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
    post:
      operationId: createSource
      tags: [sources]
      summary: Add a source to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceBody"
      responses:
        "201":
          description: Created source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
  /projects/{projectId}/sources/{id}:
    patch:
      operationId: updateSource
      tags: [sources]
      summary: Update a source configuration
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceBody"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
    delete:
      operationId: deleteSource
      tags: [sources]
      summary: Remove a source from a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/sources/{id}/trigger:
    post:
      operationId: triggerSourceCollection
      tags: [sources]
      summary: Trigger an immediate data collection run for a source
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResult"

  # ─── Signals ─────────────────────────────────────────────────────────────────
  /projects/{projectId}/signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List signals/mentions for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: sourceType
          in: query
          schema:
            type: ["string", "null"]
        - name: sentiment
          in: query
          schema:
            type: ["string", "null"]
        - name: hasSafetyFlag
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalListResponse"
  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get a single signal with full analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal with analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: Not found
  /projects/{projectId}/signals/timeline:
    get:
      operationId: getSignalTimeline
      tags: [signals]
      summary: Get signal counts over time for trending chart
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Timeline data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelinePoint"
  /projects/{projectId}/signals/safety-alerts:
    get:
      operationId: getSafetyAlerts
      tags: [signals]
      summary: Get signals flagged as safety issues
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Safety alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
  /projects/{projectId}/signals/sentiment-breakdown:
    get:
      operationId: getSentimentBreakdown
      tags: [signals]
      summary: Get sentiment breakdown for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Sentiment breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentimentBreakdown"
  /projects/{projectId}/signals/top-entities:
    get:
      operationId: getTopEntities
      tags: [signals]
      summary: Get most-mentioned entities in signals
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EntityCount"

  # ─── Analysis ────────────────────────────────────────────────────────────────
  /signals/{id}/analyze:
    post:
      operationId: analyzeSignal
      tags: [analysis]
      summary: Run AI analysis on a signal (sentiment, entities, safety, PII)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
  /projects/{projectId}/analyze-batch:
    post:
      operationId: analyzeProjectBatch
      tags: [analysis]
      summary: Run AI analysis on all unanalyzed signals in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchAnalysisResult"

  # ─── Engine Types (Admin) ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Source:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        engineTypeId:
          type: integer
        engineTypeName:
          type: string
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        latency:
          type: string
          enum: [realtime, daily, weekly]
        isEnabled:
          type: boolean
        lastRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - engineTypeId
        - engineTypeName
        - name
        - latency
        - isEnabled
        - createdAt
        - updatedAt
    CreateSourceBody:
      type: object
      properties:
        engineTypeId:
          type: integer
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        latency:
          type: string
          enum: [realtime, daily, weekly]
      required:
        - engineTypeId
        - name
        - latency
    UpdateSourceBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        config:
          type: ["object", "null"]
          additionalProperties: true
        latency:
          type: ["string", "null"]
          enum: [realtime, daily, weekly, null]
        isEnabled:
          type: ["boolean", "null"]
    TriggerResult:
      type: object
      properties:
        signalsCollected:
          type: integer
        message:
          type: string
      required:
        - signalsCollected
        - message

    # ─── Signal ───────────────────────────────────────────────────────────────
    Signal:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sourceId:
          type: integer
        sourceType:
          type: string
        externalId:
          type: ["string", "null"]
        content:
          type: string
        author:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
          enum: [positive, negative, neutral, null]
        sentimentScore:
          type: ["number", "null"]
        confidenceScore:
          type: ["number", "null"]
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        hasSafetyFlag:
          type: boolean
        safetyReason:
          type: ["string", "null"]
        hasPiiFlag:
          type: boolean
        piiTypes:
          type: array
          items:
            type: string
        isAnalyzed:
          type: boolean
        matchedKeywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - sourceId
        - sourceType
        - content
        - hasSafetyFlag
        - hasPiiFlag
        - isAnalyzed
        - entities
        - piiTypes
        - matchedKeywords
        - createdAt
    Entity:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
        confidence:
          type: number
      required:
        - text
        - type
        - confidence
    SignalListResponse:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - signals
        - total
        - limit
        - offset
    TimelinePoint:
      type: object
      properties:
        date:
          type: string
        total:
          type: integer
        positive:
          type: integer
        negative:
          type: integer
        neutral:
          type: integer
        safetyAlerts:
          type: integer
      required:
        - date
        - total
        - positive
        - negative
        - neutral
        - safetyAlerts
    SentimentBreakdown:
      type: object
      properties:
        positive:
          type: integer
        negative:
          type: integer
        neutral:
          type: integer
        unanalyzed:
          type: integer
        avgScore:
          type: ["number", "null"]
      required:
        - positive
        - negative
        - neutral
        - unanalyzed
    EntityCount:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
        count:
          type: integer
      required:
        - text
        - type
        - count
    BatchAnalysisResult:
      type: object
      properties:
        analyzed:
          type: integer
        errors:
          type: integer
        message:
          type: string
      required:
        - analyzed
        - errors
        - message

    # ─── Engine Type ──────────────────────────────────────────────────────────
...